pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "claude-telegram-bot"
version = "0.1.0"
description = "Telegram Chatbot with Claude Agent SDK on AWS Lambda"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Steven"}
]

dependencies = [
    "python-telegram-bot>=20.0",
    "anthropic>=0.40.0",
    "boto3>=1.34.0",
    "bcrypt>=4.0.0",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "moto>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
cdk = [
    "aws-cdk-lib>=2.100.0",
    "constructs>=10.0.0",
]
summarizer = [
    "playwright>=1.40.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W"]
